move tab listener to top level site and app
