+++ lib/utils/useTabListener/index.tsx
@@ -1,0 +1,27 @@
+import { useEffect } from "react";
+
+/**
+ * Outline manager for a11y, sets and unsets the .user-is-tabbing class on body
+ * so we can fine-tune outlines and focus throughout the app and site
+ */
+export const useTabListener = () => {
+  useEffect(() => {
+    const handleMousedown = () => {
+      document.body.removeEventListener("mousedown", handleMousedown);
+      document.body.classList.remove("user-is-tabbing");
+      document.body.addEventListener("keydown", handleKeydown);
+    };
+    const handleKeydown = (e: KeyboardEvent) => {
+      if (e.keyCode === 9) {
+        document.body.removeEventListener("keydown", handleKeydown);
+        document.body.classList.add("user-is-tabbing");
+        document.body.addEventListener("mousedown", handleMousedown);
+      }
+    };
+    document.body.addEventListener("keydown", handleKeydown);
+    return () => {
+      document.body.removeEventListener("keydown", handleKeydown);
+      document.body.removeEventListener("mousedown", handleMousedown);
+    };
+  }, []);
+};
